refactor(combobox): remove usage of Box
